refactor: restructure SDK components and remove unused services
- Updated the SDK export path in `src/index.ts` to reflect the new structure.
- Removed deprecated services and validation files to streamline the codebase.
- Added new documentation files to `.gitignore`.
- Updated package dependencies by removing unused packages.
- Refactored UI components to utilize new utility functions for better maintainability.
