+++ src/core/adapters/services/approval-service.ts
@@ -1,8 +1,8 @@
 import { BaseService } from '../core/base-service';
 import { getTokenContractAddress, extractErrorMessage, logger, isTestnetChain } from '../../utils';
-import { TOKEN_METADATA } from '../../constants';
+import { TOKEN_METADATA } from '../../../constants';
 import { parseUnits, formatUnits } from 'viem';
-import type { SUPPORTED_TOKENS, ApprovalInfo, SUPPORTED_CHAINS_IDS } from '../../types';
+import type { SUPPORTED_TOKENS, ApprovalInfo, SUPPORTED_CHAINS_IDS } from '../../../types';
 import type { ApprovalResult } from '../types/service-types';
 
 /**
