+++ src/ui/components/shared/index.ts
@@ -32,5 +32,5 @@ export {
 export { EnhancedInfoMessage } from './enhanced-info-message';
 
 // Animation components
-export { SlideTransition, useContentKey } from './slide-transition';
+export { SlideTransition } from './slide-transition';
 export { DragConstraintsProvider, useDragConstraints } from './drag-constraints';
